pyproject.toml
[project]
name = "complainceqapipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "azure-identity>=1.25.2",
    "azure-monitor-opentelemetry>=1.8.6",
    "azure-search-documents>=11.6.0",
    "azure-storage-blob>=12.28.0",
    "fastapi>=0.128.8",
    "firecrawl-py>=4.14.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.9",
    "langchain-text-splitters>=1.1.0",
    "langgraph>=1.0.8",
    "langsmith>=0.7.1",
    "opentelemetry-instrumentation-fastapi>=0.60b0",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pypdf>=6.7.0",
    "python-dotenv>=1.2.1",
    "redis>=7.1.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.46",
    "streamlit>=1.54.0",
    "uvicorn>=0.40.0",
    "yt-dlp>=2026.2.4",
]
